More css changes
Removed App.css as we weren't using it and it was effecting some of the desired changes. I then changed the input field so it wasn't in a card anymore I moved the results into the card next to the input group. Then I moved the buttons into the input form but they aren't quite correct yet...